+++ crates/teloxide/src/dispatching/dispatcher.rs
@@ -45,7 +45,6 @@ pub struct DispatcherBuilder<R, Err, Key> {
     ctrlc_handler: bool,
     distribution_f: fn(&Update) -> Option<Key>,
     worker_queue_size: usize,
-    stack_size: usize,
 }
 
 impl<R, Err, Key> DispatcherBuilder<R, Err, Key>
@@ -104,14 +103,6 @@ where
     #[must_use]
     pub fn worker_queue_size(self, size: usize) -> Self {
         Self { worker_queue_size: size, ..self }
-    }
-
-    /// Specifies the stack size available to the dispatcher.
-    ///
-    /// By default, it's 8 * 1024 * 1024 bytes (8 MiB).
-    #[must_use]
-    pub fn stack_size(self, size: usize) -> Self {
-        Self { stack_size: size, ..self }
     }
 
     /// Specifies the distribution function that decides how updates are grouped
@@ -186,7 +177,6 @@ where
             ctrlc_handler,
             distribution_f: _,
             worker_queue_size,
-            stack_size,
         } = self;
 
         DispatcherBuilder {
@@ -198,7 +188,6 @@ where
             ctrlc_handler,
             distribution_f: f,
             worker_queue_size,
-            stack_size,
         }
     }
 
@@ -214,7 +203,6 @@ where
             distribution_f,
             worker_queue_size,
             ctrlc_handler,
-            stack_size,
         } = self;
 
         // If the `ctrlc_handler` feature is not enabled, don't emit a warning.
@@ -229,7 +217,6 @@ where
             state: ShutdownToken::new(),
             distribution_f,
             worker_queue_size,
-            stack_size,
             workers: HashMap::new(),
             default_worker: None,
             current_number_of_active_workers: Default::default(),
@@ -270,7 +257,6 @@ pub struct Dispatcher<R, Err, Key> {
 
     distribution_f: fn(&Update) -> Option<Key>,
     worker_queue_size: usize,
-    stack_size: usize,
     current_number_of_active_workers: Arc<AtomicU32>,
     max_number_of_active_workers: Arc<AtomicU32>,
     // Tokio TX channel parts associated with chat IDs that consume updates sequentially.
@@ -310,7 +296,6 @@ where
         Err: Debug,
     {
         const DEFAULT_WORKER_QUEUE_SIZE: usize = 64;
-        const DEFAULT_STACK_SIZE: usize = 8 * 1024 * 1024;
 
         DispatcherBuilder {
             bot,
@@ -324,7 +309,6 @@ where
             ctrlc_handler: false,
             worker_queue_size: DEFAULT_WORKER_QUEUE_SIZE,
             distribution_f: default_distribution_function,
-            stack_size: DEFAULT_STACK_SIZE,
         }
     }
 }
@@ -407,28 +391,8 @@ where
         update_listener.hint_allowed_updates(&mut allowed_updates.into_iter());
 
         let stop_token = Some(update_listener.stop_token());
+        self.start_listening(update_listener, update_listener_error_handler, stop_token).await;
 
-        // We create a new Tokio runtime in order to set the correct stack size. We do
-        // it a scoped thread because Tokio runtimes cannot be nested. We need a scoped
-        // thread because of the lifetime `'a` in `&'a mut self` and because scoped
-        // threads are automatically joined. See this issue:
-        // <https://github.com/teloxide/teloxide/issues/1154>.
-        std::thread::scope(|scope: &std::thread::Scope<'_, '_>| {
-            scope.spawn(move || {
-                let runtime = tokio::runtime::Builder::new_multi_thread()
-                    .thread_stack_size(self.stack_size)
-                    .thread_name("teloxide-dispatcher-worker")
-                    .enable_all()
-                    .build()
-                    .unwrap();
-
-                runtime.block_on(self.start_listening(
-                    update_listener,
-                    update_listener_error_handler,
-                    stop_token,
-                ));
-            });
-        });
         Ok(())
     }