+++ CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## unreleased
 
+### Fixed
+
+- Stack overflow issues using `stacker` crate
+- Thread being locked because of thread spawining in dispatcher
+
+### Removed
+
+- `stack_size` setter to `DispatcherBuilder` [**BC**]
+
 ### Added
 
 - Add two examples to demonstrate how to execute functions _before_ and _after_ some endpoint: