+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## unreleased
 
+## 0.15.0 - 2025-04-04
+
 ### Added
 
 - Add two examples to demonstrate how to execute functions _before_ and _after_ some endpoint: